cleanup: reorder some cli code
