+++ cli/Makefile.autotools.am
@@ -9,7 +9,7 @@ endif
 AM_CFLAGS = -DVGMSTREAM_VERSION_AUTO -DVGM_LOG_OUTPUT -I$(top_builddir) -I$(top_srcdir) -I$(top_srcdir)/ext_includes/ $(AO_CFLAGS)
 AM_MAKEFLAGS = -f Makefile.autotools
 
-vgmstream_cli_SOURCES = vgmstream_cli.c wav_utils.c
+vgmstream_cli_SOURCES = vgmstream_cli.c vgmstream_cli_utils.c wav_utils.c
 vgmstream_cli_LDADD   = ../src/libvgmstream.la
 
 vgmstream123_SOURCES = vgmstream123.c wav_utils.c
